+++ docs/source/releasenotes.rst
@@ -21,6 +21,19 @@ Release notes
 `Released <https://pypi.org/project/rpaframework/#history>`_
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
+14.1.0 - 05 May 2022
+--------------------
+
+- Library **RPA.Robocorp.WorkItems** (:issue:`485`): Automatically parse into
+  ``email[body]`` payload variable the e-mail body on e-mail Process triggering with
+  "Parse email" configuration option enabled in Control Room.
+- Library **RPA.Hubspot** (:pr:`484`): Add keywords for creating and updating objects in
+  Hubspot, as well as a new batch system when creating batched inputs via keyword.
+- Library **RPA.Excel.Files**:
+
+  - Fix IndexError when removing .xls worksheets
+  - Fix removing currently active worksheet
+
 14.0.0 - 02 May 2022
 --------------------